refactor: extract utility functions to src/utils.js
